requirements.txt
streamlit>=1.28
langchain>=0.3.19
langchain-openai>=0.3.7
streamlit-feedback==0.1.4
load-dotenv==0.1.0
yfinance==0.2.54
pytest>=8.3.5